Send commands one character at a time.
This commit updates sendCommand to send one character at a time, because
otherwise commands in an smux file like the following will have spaces
collapsed.

echo 'echo This is pane 1' | tmux load-buffer -b PWX -

Additionally, by passing characters literally with `-l`, it avoids
unexpected magic translation of keys that tmux recognizes, such as
Enter.  This change might break existing scripts that rely on sequences
like Enter to be recognized by `tmux send-keys`, but it is hoped that
nobody is using this functionality since it was never advertised.