Attach session after joining threads.
Prior to this commit, smux tried to join the threads starting the new
session *after* attaching the session. This caused contention for the
tty on OSX between the threads trying to run tmux commands and the
already attached tmux window, breaking smux when the session was started
outside a tmux session.

This commit fixes the problem by always attaching last.