Use hex escape codes for send-keys.
Previously, smux used `send-keys -l` and did some nasty quote
manipulation on the commands to ensure that they were sent correctly
even when they contained special characters such as single quotes.

This approach worked for a long time, but it does not handle commands
that start with `-`, because the tmux cli will treat them as options and
try to interpret them, instead of treating them as characters to send.

This commit changes smux to eliminate the nasty quote manipulation and
use `send-keys -H` with hex versions of each character instead.

Unfortunately, the `-H` flag does not handle unicode, so this will only
work for ASCII-only commands.