Enable leading dashes in command lines.
Previously, leading dashes in an smux file did not work correctly because the
tmux cli considered them to be options instead of a literal string to
pass.

This commit adds a '' in front of the keys passed to `send-keys`. This
forced them to be interpreted literally.

This idea was suggested here:
https://unix.stackexchange.com/a/472112/34334